+++ lib/remote.d.ts
@@ -1,4 +1,4 @@
-import { _ as readExpertPrompt, v as readLocalizedExpertPrompt } from "./index-BXQQzToz.js";
+import { _ as readExpertPrompt, v as readLocalizedExpertPrompt } from "./index-D4dhi1A4.js";
 import { z } from "zod";
 import { TypertRemoteService } from "@deepseek-ai/dsh-typert-protocol";
 import { Context } from "@deepseek-ai/cordis";